+++ backend/src/main/java/com/staccato/member/controller/MemberController.java
@@ -4,14 +4,12 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import com.staccato.config.auth.LoginMember;
+import com.staccato.member.controller.docs.MemberControllerDocs;
 import com.staccato.member.domain.Member;
 import com.staccato.member.service.MemberService;
 import com.staccato.member.service.dto.request.MemberReadRequest;
-import com.staccato.member.service.dto.response.MemberResponses;
-import com.staccato.member.service.dto.response.MemberSearchResponse;
 import com.staccato.member.service.dto.response.MemberSearchResponses;
 import lombok.RequiredArgsConstructor;
