+++ backend/src/main/java/com/staccato/member/controller/docs/MemberControllerDocs.java
@@ -1,4 +1,4 @@
-package com.staccato.member.controller;
+package com.staccato.member.controller.docs;
 
 import org.springframework.http.ResponseEntity;
 import com.staccato.config.auth.LoginMember;
